+++ apps/api/src/mcp/tools/update-alert-rule.ts
@@ -10,6 +10,7 @@ import { Effect, Option, Schema } from "effect"
 import { createDualContent } from "@/mcp/lib/structured-output"
 import { resolveTenant } from "@/mcp/lib/query-warehouse"
 import { AlertsService } from "@/services/alerts/AlertsService"
+import { AlertRulesService } from "@/services/alerts/AlertRulesService"
 import { AlertRuleUpsertRequest, type AlertRuleDocument } from "@maple/domain/http"
 
 const decodeAlertRuleRequest = Schema.decodeUnknownEffect(AlertRuleUpsertRequest)
@@ -192,8 +193,9 @@ export function registerUpdateAlertRuleTool(server: McpToolRegistrar) {
 		Effect.fn("McpTool.updateAlertRule")(function* (params) {
 			const tenant = yield* resolveTenant
 			const alerts = yield* AlertsService
+			const rules = yield* AlertRulesService
 
-			const list = yield* alerts.listRules(tenant.orgId).pipe(
+			const list = yield* rules.listRules(tenant.orgId).pipe(
 				Effect.mapError(
 					(error) =>
 						new McpQueryError({
