+++ apps/api/src/routes/v2/alerts.http.test.ts
@@ -19,6 +19,7 @@ import { DashboardPersistenceService } from "@/services/dashboards/DashboardPers
 import { AlertRuntime, AlertsService } from "@/services/alerts/AlertsService"
 import { AlertDestinationsService } from "@/services/alerts/AlertDestinationsService"
 import { AlertReadModelsService } from "@/services/alerts/AlertReadModelsService"
+import { AlertRulesService } from "@/services/alerts/AlertRulesService"
 import { HazelOAuthService } from "@/services/auth/HazelOAuthService"
 import { OrgClickHouseSettingsService } from "@/services/org/OrgClickHouseSettingsService"
 import { OrgMembersService } from "@/services/org/OrgMembersService"
@@ -110,6 +111,9 @@ const makeHarness = (warehouseService: WarehouseQueryServiceShape = warehouseStu
 	const alertReadModelsLive = AlertReadModelsService.layer.pipe(
 		Layer.provide(Layer.mergeAll(testDb.layer, warehouseLive)),
 	)
+	const alertRulesLive = AlertRulesService.layer.pipe(
+		Layer.provide(Layer.mergeAll(testDb.layer, runtimeLive)),
+	)
 	const alertsLive = AlertsService.layer.pipe(
 		Layer.provide(
 			Layer.mergeAll(
@@ -125,6 +129,7 @@ const makeHarness = (warehouseService: WarehouseQueryServiceShape = warehouseStu
 				investigationsLive,
 				alertDestinationsLive,
 				alertReadModelsLive,
+				alertRulesLive,
 			),
 		),
 	)
@@ -134,6 +139,7 @@ const makeHarness = (warehouseService: WarehouseQueryServiceShape = warehouseStu
 		DashboardPersistenceService.layer,
 		alertDestinationsLive,
 		alertReadModelsLive,
+		alertRulesLive,
 		alertsLive,
 	).pipe(Layer.provideMerge(Layer.mergeAll(envLive, testDb.layer)))
