+++ apps/api/src/runtime/graph-boundaries.test.ts
@@ -69,6 +69,7 @@ describe("API runtime graph boundaries", () => {
 
 		expect(layerMembers(source, "McpRuntimeServicesLive")).toEqual([
 			"AlertReadModelsServiceLive",
+			"AlertRulesServiceLive",
 			"AlertsServiceLive",
 			"DashboardPersistenceService.layer",
 			"ErrorActorsServiceLive",
